move topl-android classes to new directory
